+++ src/main/java/com/assemblyrobot/simulator/system/components/MaterialQueue.java
@@ -1,4 +1,4 @@
-package com.assemblyrobot.system.core;
+package com.assemblyrobot.simulator.system.components;
 
 import java.util.LinkedList;
 import lombok.Getter;
