+++ src/main/java/com/assemblyrobot/simulator/system/engines/MainEngine.java
@@ -1,9 +1,7 @@
-package com.assemblyrobot.system.engines;
+package com.assemblyrobot.simulator.system.engines;
 
 import com.assemblyrobot.simulator.core.Engine;
-import com.assemblyrobot.system.generators.ArrivalEventPropagator;
-import com.assemblyrobot.system.stations.AssemblyStation;
-import java.util.Arrays;
+import com.assemblyrobot.simulator.system.generators.ArrivalEventPropagator;
 
 public class MainEngine extends Engine {
   private final ArrivalEventPropagator arrivalEventPropagator =
