+++ src/main/java/com/assemblyrobot/simulator/system/stations/AssemblyStation.java
@@ -1,8 +1,7 @@
-package com.assemblyrobot.system.stations;
+package com.assemblyrobot.simulator.system.stations;
 
-import com.assemblyrobot.simulator.core.events.EventQueue;
 import com.assemblyrobot.simulator.core.generators.AssemblyTimeGenerator;
-import com.assemblyrobot.system.core.Station;
+import com.assemblyrobot.simulator.system.components.Station;
 
 public class AssemblyStation extends Station {
